+++ types/fastify.d.ts
@@ -1,9 +1,0 @@
-declare module "fastify" {
-  const fastify: any;
-  export default fastify;
-}
-
-declare module "@fastify/websocket" {
-  const websocket: any;
-  export default websocket;
-}